google_sheets_connector/requirements.txt
# Google Sheets API dependencies
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Data manipulation
pandas>=2.0.0
numpy>=1.21.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Date and time utilities
python-dateutil>=2.8.0

# Logging and monitoring
colorlog>=6.0.0

# JSON handling
jsonschema>=4.0.0
